% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm-data.R
\docType{data}
\name{fm}
\alias{fm}
\title{Dado: matriz numerica de acordo com um filtro para o focal}
\format{
matrix
}
\source{
Valores aleatórios
}
\usage{
fm
}
\description{
Uma matriz com valores de 0 e 1 com 7 colunas e 7 linhas.
O filtro considerado é type="circle", então os valores de 1 estão
distribuidos de forma circular, enquanto
os valores de 0 estão ao redor do círculo.
}
\keyword{datasets}
