% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-data.R
\docType{data}
\name{info}
\alias{info}
\title{Informações sobre "map"}
\format{
dataframe
}
\source{
Valores aleatórios
}
\usage{
info
}
\description{
Um data frame com informações quanto ao número de elementos da paisagem
considerados em "map", quais os nomes destes elementos e seus
respectivos pesos
}
\keyword{datasets}
