% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-quality.R
\name{quality}
\alias{quality}
\title{Calcula a qualidade da matriz}
\usage{
quality(m, fm, lvl, q, na.rm)
}
\arguments{
\item{m}{SpatRaster original}

\item{fm}{Objeto de matriz numérica para o "move window" do focal}

\item{lvl}{Número do nível (ou elemento da matriz) que será calculado a qualidade}

\item{q}{Valor de peso para calcular a qualidade de lvl}

\item{na.rm}{Argumento lógico}
}
\value{
Um SpatRaster
}
\description{
Calcula a qualidade da matriz
}
\examples{
mm <- base::matrix(base::sample(1:10,
                   size = 100, replace = TRUE), ncol = 10, nrow = 10)
map <- terra::rast(mm)

fm <- base::matrix(c(0, 0, 1, 1, 1, 0, 0,
                     0, 1, 1, 1, 1, 1, 0,
                     1, 1, 1, 1, 1, 1, 1,
                     1, 1, 1, 1, 1, 1, 1,
                     1, 1, 1, 1, 1, 1, 1,
                     0, 1, 1, 1, 1, 1, 0,
                     0, 0, 1, 1, 1, 0, 0),
                   ncol = 7, nrow = 7)

info <- base::data.frame(lvl_cod = c(1:10),
                         habitat = c("Floresta madura", "Floresta incial",
                                     "Savana", "Pastagem",
                                     "Agricultura", "Silvicultura",
                                     "N-vegetada", "Urbano",
                                     "Mineiracao", "Agua"),
                          q_peso = c(10:1))

m <- map
fm <- fm
lvl <- info$lvl_cod[1]
q <- info$q_peso[1]
na.rm <- TRUE

quality(m, fm, lvl, q, na.rm)

}
