% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm-data.R
\docType{data}
\name{fm}
\alias{fm}
\title{Matriz circular}
\format{
matrix
}
\source{
Valores aleatórios
}
\usage{
fm
}
\description{
Uma matriz com valores de 0 e 1 com 7 colunas e 7 linhas.
Os valores de 1 estão distribuidos de forma circular, enquanto
os valores de 0 estão ao redor do círculo.
}
\keyword{datasets}
