% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-data.R
\docType{data}
\name{info}
\alias{info}
\title{Informacoes hipoteticas para detalhar um SpatRaster}
\format{
dataframe
}
\source{
Valores aleatórios
}
\usage{
info
}
\description{
Dataframe com 10 linhas e 3 colunas.
Coluna 'lvl_cod' representa os códigos de cada nível de elemento da paisagem.
Coluna 'habitat' nomeia esses elementos de acordo com o habitat representado.
Coluna 'q_peso' atribui um peso para cada um dos elementos.
}
\keyword{datasets}
