% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-mx_weight.R
\name{mx_weight}
\alias{mx_weight}
\title{Produz uma matriz para o focal}
\usage{
mx_weight(m, tam, pix, type)
}
\arguments{
\item{m}{SpatRaster original}

\item{tam}{Tamanho desejado da janela do focal}

\item{pix}{Tamanho do pixel do SpatRaster}

\item{type}{Tipo de filtro que será aplicado para criar a matriz
Tipos podem ser circular ("circle) ou retanular ("rectangle)}
}
\value{
Uma matriz
}
\description{
Produz uma matriz para o focal
}
\examples{
 mm <- base::matrix(base::sample(1:10,
                    size = 100, replace = TRUE), ncol = 10, nrow = 10)
 map <- terra::rast(mm)

 m <- map
 tam <- 100
 pix <- 30
 type <- "circle"

 mx_weight(m, tam, pix, type)

}
