% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-quality.R
\name{quality}
\alias{quality}
\title{Calculo da qualidade da matriz}
\usage{
quality(m, fm, lvl, q, na.rm, tpeso)
}
\arguments{
\item{m}{Objeto SpatRaster original}

\item{fm}{Objeto de matriz numérica que será utilizado no "move window" do focal}

\item{lvl}{Código do nível (elemento da matriz) que será calculado a qualidade}

\item{q}{Valor de peso para calcular a qualidade de 'lvl'}

\item{na.rm}{Argumento lógico}

\item{tpeso}{Valor da soma dos pesos}
}
\value{
Um objeto SpatRaster
}
\description{
Possibilita o cálculo de quão presente está um determinado elemento na
paisagem ponderado pelo peso que definido para o mesmo.
}
\examples{
mm <- base::matrix(base::sample(1:10,
                   size = 100, replace = TRUE), ncol = 10, nrow = 10)
map <- terra::rast(mm)

m <- map
fm <- fm
lvl <- info$lvl_cod[1]
q <- info$q_peso[1]
na.rm <- TRUE
tpeso <- sum(info$q_peso)

quality(m, fm, lvl, q, na.rm, tpeso)

}
