% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual.R
\name{quality}
\alias{quality}
\title{Calcula a qualidade da matriz}
\usage{
quality(m, fm, lvl, q, na.rm, hab)
}
\arguments{
\item{m}{SpatRaster original}

\item{fm}{Objeto de matriz numérica para o "move window" do focal}

\item{lvl}{Número do nível (ou elemento da matriz) que será calculado a qualidade}

\item{q}{Valor de peso para calcular a qualidade de lvl}

\item{na.rm}{Argumento lógico}

\item{hab}{Argumento do tipo "character" que representa o nome do habitat
que sera considerado em "lvl}
}
\value{
Um SpatRaster
}
\description{
Calcula a qualidade da matriz
}
\examples{
map <- mapa_teste

map2 <- base::matrix(c(1,1, # floresta
                       3,1,
                       4,1,
                       5,1,
                      49,1,
                      10,2, # natural nao-florestal
                      11,2,
                      12,2,
                      32,2,
                      29,2,
                      13,2,
                      14,3, # pastagem
                      15,3,
                      18,4, # agricultura
                      19,4,
                      39,4,
                      20,4,
                      40,4,
                      41,4,
                      36,4,
                      46,4,
                      47,4,
                      48,4,
                       9,5, # silvicultura
                      21,3,
                      22,6, # nao-vegetada
                      23,6,
                      24,7, # urbano
                      30,8, # mineira??o
                      25,6,
                      26,9, # agua
                      33,9,
                      31,9,
                      27,7), ncol = 2, byrow = TRUE)

map_C <- terra::classify(map, map2)

fm <- maqual::mx_weight(map_c, 100, 30, "circle")

info <- base::data.frame(lvl_cod = c(1:9),
                         habitat = c("Floresta", "Natural n-florestal", "Pastagem",
                                     "Agricultura", "Silvicultura", "N-vegetada",
                                     "Urbano", "Mineiracao", "Agua"),
                          q_peso = c(9:1))

m <- map_c
fm <- fm
lvl <- info$lvl_cod[1]
q <- info$q_peso[1]
na.rm <- TRUE
hab <- info$habitat[1]

}
