% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-quality.R
\name{quality}
\alias{quality}
\title{Calcula a qualidade da matriz}
\usage{
quality(m, fm, lvl, q, na.rm)
}
\arguments{
\item{m}{SpatRaster original}

\item{fm}{Objeto de matriz numérica para o "move window" do focal}

\item{lvl}{Número do nível (ou elemento da matriz) que será calculado a qualidade}

\item{q}{Valor de peso para calcular a qualidade de lvl}

\item{na.rm}{Argumento lógico}
}
\value{
Um SpatRaster
}
\description{
Calcula a qualidade da matriz
}
\examples{
mm <- base::matrix(base::sample(1:10,
                   size = 100, replace = TRUE), ncol = 10, nrow = 10)
map <- terra::rast(mm)

m <- map
fm <- fm
lvl <- info$lvl_cod[1]
q <- info$q_peso[1]
na.rm <- TRUE

quality(m, fm, lvl, q, na.rm)

}
